+++ include/kyosu/functions/digamma.hpp
@@ -8,22 +8,18 @@
 #pragma once
 
 #include <kyosu/details/invoke.hpp>
-#include <eve/module/math.hpp>
+#include <eve/module/special.hpp>
 
 namespace kyosu::tags
 {
-  struct callable_faddeeva : eve::elementwise
+  struct callable_digamma : eve::elementwise
   {
-    using callable_tag_type = callable_faddeeva;
+    using callable_tag_type = callable_digamma;
 
-    KYOSU_DEFERS_CALLABLE(faddeeva_);
+    KYOSU_DEFERS_CALLABLE(digamma_);
 
     template<eve::ordered_value T>
-    static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v) noexcept
-    {
-      auto fn = callable_faddeeva{};
-      return fn(complex(v));
-    }
+    static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v) noexcept { return eve::digamma(v); }
 
     template<typename T>
     KYOSU_FORCEINLINE auto operator()(T const& target) const noexcept -> decltype(eve::tag_invoke(*this, target))
@@ -32,7 +28,7 @@ namespace kyosu::tags
     }
 
     template<typename... T>
-    eve::unsupported_call<callable_faddeeva(T&&...)> operator()(T&&... x) const
+    eve::unsupported_call<callable_digamma(T&&...)> operator()(T&&... x) const
     requires(!requires { eve::tag_invoke(*this, KYOSU_FWD(x)...); }) = delete;
   };
 }
@@ -42,8 +38,8 @@ namespace kyosu
 //======================================================================================================================
 //! @addtogroup functions
 //! @{
-//!   @var faddeeva
-//!   @brief  Callable object computing \f$e^{-z^2}\mathrm{erfc}(-iz)\f$ the scaled complex error func
+//!   @var digamma
+//!   @brief Computes the Digamma function i.e. the logarithmic derivative of the \f$\Gamma\f$ function
 //!
 //!   **Defined in Header**
 //!
@@ -56,23 +52,22 @@ namespace kyosu
 //!   @code
 //!   namespace kyosu
 //!   {
-//!      template<eve::ordered_value T>       constexpr auto faddeeva(T z) noexcept;  //1
-//!      template<kyosu::concepts::complex T> constexpr auto faddeeva(T z) noexcept;  //2
+//!     constexpr auto  digamma(T z) noexcept;
 //!   }
 //!   @endcode
 //!
 //!   **Parameters**
 //!
-//!     * `z` : Value to process.
+//!     * `z` : value to process.
 //!
-//! **Return value**
+//!   **Return value**
 //!
-//!   Returns \f$e^{-z^2}\mathrm{erfc}(-iz)\f$ the scaled complex error function
+//!     The value of the Digamma function: \f$\frac{\Gamma'(z)}{\Gamma(z)}\f$ is returned.
 //!
 //!  @groupheader{Example}
 //!
-//!  @godbolt{doc/faddeeva.cpp}
+//!  @godbolt{doc/digamma.cpp}
 //! @}
 //======================================================================================================================
-inline constexpr tags::callable_faddeeva faddeeva = {};
+inline constexpr tags::callable_digamma digamma = {};
 }
