+++ include/kyosu/functions/erfcx.hpp
@@ -12,27 +12,26 @@
 
 namespace kyosu::tags
 {
-  struct callable_lambda : eve::elementwise
+  struct callable_erfcx : eve::elementwise
   {
-    using callable_tag_type = callable_lambda;
+    using callable_tag_type = callable_erfcx;
 
-    KYOSU_DEFERS_CALLABLE(lambda_);
+    KYOSU_DEFERS_CALLABLE(erfcx_);
 
     template<eve::ordered_value T>
     static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v) noexcept
     {
-      auto fn = callable_lambda{};
-      return fn(complex(v));
+      return eve::erfcx(v);
     }
 
     template<typename T>
-    KYOSU_FORCEINLINE auto operator()(T const & target) const noexcept -> decltype(eve::tag_invoke(*this, target))
+    KYOSU_FORCEINLINE auto operator()(T const& target) const noexcept -> decltype(eve::tag_invoke(*this, target))
     {
       return eve::tag_invoke(*this, target);
     }
 
     template<typename... T>
-    eve::unsupported_call<callable_lambda(T&&...)> operator()(T&&... x) const
+    eve::unsupported_call<callable_erfcx(T&&...)> operator()(T&&... x) const
     requires(!requires { eve::tag_invoke(*this, KYOSU_FWD(x)...); }) = delete;
   };
 }
@@ -42,12 +41,9 @@ namespace kyosu
 //======================================================================================================================
 //! @addtogroup functions
 //! @{
-//!   @var lambda
-//! @brief Callable object computing The Dirichlet \f$ \displaystyle \lambda(z) = \sum_0^\infty \frac{1}{(2n+1)^z}\f$
-//!
-//! This function can be extended to the whole complex plane as \f$\lambda(z) = \zeta(z)(1-2^{-x})\f$
-//! (where \f$\zeta\f$ is the Riemann zeta function). It coincides with the serie where the serie converges.
-//! However for `z = 1` the result is \f$\infty\f$.
+//!   @var erfcx
+//!   @brief Computes the normalized complementary error function
+//!   \f$ \displaystyle \mbox{erfcx}(x) = e^{x^2} \mbox{erfc}(x)\f$.
 //!
 //!   **Defined in Header**
 //!
@@ -60,23 +56,25 @@ namespace kyosu
 //!   @code
 //!   namespace kyosu
 //!   {
-//!      template<unsigned_scalar_value K, eve::ordered_value T>       constexpr auto lambda(K, k, T z) noexcept;  //1
-//!      template<unsigned_scalar_value K, kyosu::concepts::complex T> constexpr auto lambda(K, k, T z) noexcept;  //2
+//!      template<eve::ordered_value T>              constexpr auto erfcx(T z) noexcept;  //1
+//!      template<kyosu::concepts::cayley_dickson T> constexpr auto erfcx(T z) noexcept;  //2
 //!   }
 //!   @endcode
 //!
 //!   **Parameters**
 //!
-//!     * `z` : complex or real value to process.
+//!     * `z` : Value to process.
 //!
 //! **Return value**
 //!
-//!    Returns the Dirichlet sum \f$  \displaystyle \sum_0^\infty \frac{1}{(2n+1)^z}\f$
+//!   1. a real input z return eve::erfcx(z).
+//!
+//!   2. The value of the normalized complementary error function is returned.
 //!
 //!  @groupheader{Example}
 //!
-//!  @godbolt{doc/lambda.cpp}
+//!  @godbolt{doc/erfcx.cpp}
 //! @}
 //======================================================================================================================
-inline constexpr tags::callable_lambda lambda = {};
+inline constexpr tags::callable_erfcx erfcx = {};
 }
