+++ include/kyosu/functions/lbeta.hpp
@@ -8,27 +8,28 @@
 #pragma once
 
 #include <kyosu/details/invoke.hpp>
-#include <eve/module/special.hpp>
 
 namespace kyosu::tags
 {
-  struct callable_digamma : eve::elementwise
+  struct callable_lbeta : eve::elementwise
   {
-    using callable_tag_type = callable_digamma;
+    using callable_tag_type = callable_lbeta;
 
-    KYOSU_DEFERS_CALLABLE(digamma_);
+    KYOSU_DEFERS_CALLABLE(lbeta_);
 
-    template<eve::ordered_value T>
-    static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v) noexcept { return eve::digamma(v); }
+    template<eve::ordered_value T, eve::ordered_value U>
+    static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v, U const& w) noexcept {
+      auto fn = callable_lbeta{};
+      return fn(complex(v), w); }
 
-    template<typename T>
-    KYOSU_FORCEINLINE auto operator()(T const& target) const noexcept -> decltype(eve::tag_invoke(*this, target))
+    template<typename T, typename U>
+    KYOSU_FORCEINLINE auto operator()(T const& target1, U const& target2) const noexcept -> decltype(eve::tag_invoke(*this, target1, target2))
     {
-      return eve::tag_invoke(*this, target);
+      return eve::tag_invoke(*this, target1, target2);
     }
 
     template<typename... T>
-    eve::unsupported_call<callable_digamma(T&&...)> operator()(T&&... x) const
+    eve::unsupported_call<callable_lbeta(T&&...)> operator()(T&&... x) const
     requires(!requires { eve::tag_invoke(*this, KYOSU_FWD(x)...); }) = delete;
   };
 }
@@ -38,8 +39,8 @@ namespace kyosu
 //======================================================================================================================
 //! @addtogroup functions
 //! @{
-//!   @var digamma
-//!   @brief Computes the Digamma function i.e. the logarithmic derivative of the \f$\Gamma\f$ function
+//!   @var lbeta
+//!   @brief Computes the natural logarithm of the lbeta function.
 //!
 //!   **Defined in Header**
 //!
@@ -52,23 +53,22 @@ namespace kyosu
 //!   @code
 //!   namespace kyosu
 //!   {
-//!      template<kyosu::concepts::complex T>   constexpr T  digamma(T z) noexcept;
-//!      template<eve::floating_ordered_value T> constexpr T  digamma(T z) noexcept;
+//!      auto lbeta(auto x,auto y) noexcept;
 //!   }
 //!   @endcode
 //!
 //!   **Parameters**
 //!
-//!     * `z` : real or complex value to process.
+//!     * `x`,`y` : Values to process.
 //!
 //!   **Return value**
 //!
-//!     The value of the Digamma function: \f$\frac{\Gamma'(z)}{\Gamma(z)}\f$ is returned.
+//!    `log(beta(x, y)`.
 //!
 //!  @groupheader{Example}
 //!
-//!  @godbolt{doc/digamma.cpp}
+//!  @godbolt{doc/lbeta.cpp}
 //! @}
 //======================================================================================================================
-inline constexpr tags::callable_digamma digamma = {};
+inline constexpr tags::callable_lbeta lbeta = {};
 }
