+++ include/kyosu/functions/tgamma.hpp
@@ -8,31 +8,29 @@
 #pragma once
 
 #include <kyosu/details/invoke.hpp>
-#include <eve/module/math.hpp>
 
 namespace kyosu::tags
 {
-  struct callable_zeta : eve::elementwise
+  struct callable_tgamma : eve::elementwise
   {
-    using callable_tag_type = callable_zeta;
+    using callable_tag_type = callable_tgamma;
 
-    KYOSU_DEFERS_CALLABLE(zeta_);
+    KYOSU_DEFERS_CALLABLE(tgamma_);
 
     template<eve::ordered_value T>
-    static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v) noexcept
-    {
-      auto fn = callable_zeta{};
+    static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v) noexcept {
+      auto fn = callable_tgamma{};
       return fn(complex(v));
     }
 
     template<typename T>
-    KYOSU_FORCEINLINE auto operator()(T const & target) const noexcept -> decltype(eve::tag_invoke(*this, target))
+    KYOSU_FORCEINLINE auto operator()(T const& target) const noexcept -> decltype(eve::tag_invoke(*this, target))
     {
       return eve::tag_invoke(*this, target);
     }
 
     template<typename... T>
-    eve::unsupported_call<callable_zeta(T&&...)> operator()(T&&... x) const
+    eve::unsupported_call<callable_tgamma(T&&...)> operator()(T&&... x) const
     requires(!requires { eve::tag_invoke(*this, KYOSU_FWD(x)...); }) = delete;
   };
 }
@@ -42,8 +40,8 @@ namespace kyosu
 //======================================================================================================================
 //! @addtogroup functions
 //! @{
-//!   @var zeta
-//!   @brief Computes the Riemann \f$ \displaystyle\zeta(z)=\sum_0^\infty \frac{1}{(n+1)^z}\f$.
+//!   @var tgamma
+//!   @brief Computes \f$\Gamma(z)\f$r.
 //!
 //!   **Defined in Header**
 //!
@@ -56,23 +54,22 @@ namespace kyosu
 //!   @code
 //!   namespace kyosu
 //!   {
-//!      template<eve::floating_ordered_value T> constexpr auto zeta(T z) noexcept;
-//!      template<kyosu::concepts::complex T>  constexpr auto zeta(T z) noexcept;
+//!      constexpr T  tgamma(T z) noexcept;
 //!   }
 //!   @endcode
 //!
 //!   **Parameters**
 //!
-//!     * `z` : complex or real value to process.
+//!     * `z` : Value to process.
 //!
-//! **Return value**
+//!   **Return value**
 //!
-//!   Returns the Dirichlet zeta function: \f$  \displaystyle \sum_0^\infty \frac{1}{(n+1)^z}\f$
+//!     Returns \f$\Gamma(z)\f$.
 //!
 //!  @groupheader{Example}
 //!
-//!  @godbolt{doc/zeta.cpp}
+//!  @godbolt{doc/tgamma.cpp}
 //! @}
 //======================================================================================================================
-inline constexpr tags::callable_zeta zeta = {};
+inline constexpr tags::callable_tgamma tgamma = {};
 }
