+++ include/kyosu/functions/zeta.hpp
@@ -12,27 +12,27 @@
 
 namespace kyosu::tags
 {
-  struct callable_deta : eve::elementwise
+  struct callable_zeta : eve::elementwise
   {
-    using callable_tag_type = callable_deta;
+    using callable_tag_type = callable_zeta;
 
-    KYOSU_DEFERS_CALLABLE(deta_);
+    KYOSU_DEFERS_CALLABLE(zeta_);
 
-    template<eve::unsigned_scalar_value K, eve::ordered_value T>
-    static KYOSU_FORCEINLINE auto deferred_call(auto, K const & k, T const& v) noexcept
+    template<eve::ordered_value T>
+    static KYOSU_FORCEINLINE auto deferred_call(auto, T const& v) noexcept
     {
-      auto fn = callable_deta{};
-      return fn(k, complex(v));
+      auto fn = callable_zeta{};
+      return fn(complex(v));
     }
 
-    template<typename K, typename T>
-    KYOSU_FORCEINLINE auto operator()(K const & target1, T const& target2) const noexcept -> decltype(eve::tag_invoke(*this, target1, target2))
+    template<typename T>
+    KYOSU_FORCEINLINE auto operator()(T const & target) const noexcept -> decltype(eve::tag_invoke(*this, target))
     {
-      return eve::tag_invoke(*this, target1, target2);
+      return eve::tag_invoke(*this, target);
     }
 
     template<typename... T>
-    eve::unsupported_call<callable_deta(T&&...)> operator()(T&&... x) const
+    eve::unsupported_call<callable_zeta(T&&...)> operator()(T&&... x) const
     requires(!requires { eve::tag_invoke(*this, KYOSU_FWD(x)...); }) = delete;
   };
 }
@@ -42,8 +42,8 @@ namespace kyosu
 //======================================================================================================================
 //! @addtogroup functions
 //! @{
-//!   @var deta
-//!   @brief Computes the Dirichlet sums \f$ \displaystyle \sum_{n = 0}^\infty \frac{(-1)^n}{(kn+1)^z}\f$.
+//!   @var zeta
+//!   @brief Computes the Riemann \f$ \displaystyle\zeta(z)=\sum_0^\infty \frac{1}{(n+1)^z}\f$.
 //!
 //!   **Defined in Header**
 //!
@@ -56,24 +56,22 @@ namespace kyosu
 //!   @code
 //!   namespace kyosu
 //!   {
-//!      template<unsigned_scalar_value K, eve::ordered_value T>       constexpr auto deta(K k, T z) noexcept;  //1
-//!      template<unsigned_scalar_value K, kyosu::concepts::complex T> constexpr auto deta(K k, T z) noexcept;  //2
+//!      constexpr auto zeta(auto z) noexcept;
 //!   }
 //!   @endcode
 //!
 //!   **Parameters**
 //!
-//!     * `k` : scalar unsigned value,  parameter of the sum.
 //!     * `z` : complex or real value to process.
 //!
 //! **Return value**
 //!
-//!   Returns the Dirichlet sum \f$  \displaystyle \sum_{n = 0}^\infty \frac{(-1)^n}{(kn+1)^z}\f$
+//!   Returns the Dirichlet zeta function: \f$  \displaystyle \sum_0^\infty \frac{1}{(n+1)^z}\f$
 //!
 //!  @groupheader{Example}
 //!
-//!  @godbolt{doc/deta.cpp}
+//!  @godbolt{doc/zeta.cpp}
 //! @}
 //======================================================================================================================
-inline constexpr tags::callable_deta deta = {};
+inline constexpr tags::callable_zeta zeta = {};
 }
